// src/app/ficha-de-matricula/models/matricula.ts
export class DatosComplementarios {
    matriculaId: number = 0;
    previsionDeSaludId: number = 0;
    problemasDeSalud: string = "";
    contraindicacionDeActividadFisica: boolean = false;
    tratamientoMedico: boolean = false;
    detalleDeSalud: string = "";
    procedenciaMismoColegio: boolean = false;
    nombreDelOtroColegio: string = "";
    dependenciaDelOtroColegio: string = "";
    motivoDeCambio: string = "";
}